Introduce `ErrorMessages` Type
The nested structure of error messages is difficult to manage and somewhat squishy…

Validators always return `array<string, string>`, but input filter error message arrays can be nested indefinitely - these types of recursive structures are a nightmare for static analysis and can easily be mutated unintentionally.

An object is easier to reason about, easier to process recursively in a type-safe way and easier to refactor and identify.

Additionally, we can get an accurate count of actual error messages instead of counting empty arrays - i.e. `count($inputFilter->getMessages())` will be zero, even if it contains an empty set of errors for each valid input.

Signed-off-by: George Steel <george@net-glue.co.uk>